+++ src/main.ts
@@ -1,33 +1,48 @@
+import { ValidationPipe } from "@nestjs/common"
+import { ConfigService } from "@nestjs/config"
 import { NestFactory } from "@nestjs/core"
 import { FastifyAdapter } from "@nestjs/platform-fastify"
 import type { NestFastifyApplication } from "@nestjs/platform-fastify"
 import { DocumentBuilder, SwaggerModule } from "@nestjs/swagger"
-import * as dotenv from "dotenv-flow"
 import { promises as fs } from "fs"
 import * as path from "path"
 import * as packagejson from "../package.json"
 import { AppModule } from "./app.module"
 
-dotenv.config()
-const { HOST } = process.env
-
-const bootstrap = async () => {
+async function main() {
   const app = await NestFactory.create<NestFastifyApplication>(
     AppModule,
     new FastifyAdapter()
   )
+  app.useGlobalPipes(new ValidationPipe())
 
-  const options = new DocumentBuilder()
-    .setTitle(packagejson.name)
-    .setDescription(packagejson.description)
-    .setVersion(packagejson.version)
-    .addServer(HOST ? `https://${HOST}` : "http://localhost:3000")
-    .build()
-  const document = SwaggerModule.createDocument(app, options)
-  SwaggerModule.setup("docs", app, document)
-  // export OpenAPI/Swagger specs to JSON file, in order to work with Saasify
-  await fs.writeFile(path.join(__dirname, "../openapi.json"), JSON.stringify(document))
+  const config = app.get(ConfigService)
+  const nodeEnv = config.get<string>("NODE_ENV")
+  const port = config.get<string>("PORT", "3000")
 
-  await app.listen(3000, "::")
+  if (nodeEnv !== "prod") {
+    const document = SwaggerModule.createDocument(
+      app,
+      new DocumentBuilder()
+        .setTitle(packagejson.name)
+        .setDescription(packagejson.description)
+        .setVersion(packagejson.version)
+        .addServer(`http://localhost:${parseInt(port)}`)
+        .build()
+    )
+    SwaggerModule.setup("docs", app, document)
+
+    // export OpenAPI/Swagger specs to JSON file
+    // should be dist/openapi.json
+    const filepath = path.join(__dirname, "../openapi.json")
+    fs.writeFile(filepath, JSON.stringify(document)).catch((err) => {
+      console.error(err)
+    })
+  }
+
+  await app.listen(parseInt(port), "::")
 }
-bootstrap()
+
+main().catch((err) => {
+  console.error(err)
+})