+++ src/renderer/renderer.service.ts
@@ -1,12 +1,16 @@
 import { Inject, Injectable } from "@nestjs/common"
+import axios from "axios"
 import { firefox } from "playwright-firefox"
 import type { FirefoxBrowser } from "playwright-firefox"
-import { BlocklistProvider } from "../blocklist/blocklist.provider"
 import { HttpProxy } from "../common/types/http-proxy.class"
 import { HeadlessBrowserProvider } from "../headless-browser/headless-browser.provider"
+import { BlocklistProvider } from "../resources/blocklist.provider"
 
 @Injectable()
 export class RendererService {
+  // temporarily use this axios instance until HttpService uses axios@0.20.x internally
+  axios = axios.create({ timeout: 10000, validateStatus: () => true })
+
   public constructor(
     @Inject(BlocklistProvider.providerName) private readonly blocklist: Set<string>,
     @Inject(HeadlessBrowserProvider.providerName) private readonly browser: FirefoxBrowser
@@ -18,25 +22,19 @@ export class RendererService {
     headers?: Record<string, string>,
     proxy?: HttpProxy
   ): Promise<string> {
-    let proxyBrowser
-    let context
-    if (proxy) {
-      proxyBrowser = await firefox.launch({
-        proxy: {
-          server: `${proxy.host}:${proxy.port}`,
-          username: proxy.username,
-          password: proxy.password,
-        },
-      })
-      context = await proxyBrowser.newContext({
-        extraHTTPHeaders: headers,
-      })
-    } else {
-      context = await this.browser.newContext({
-        extraHTTPHeaders: headers,
-      })
-    }
-    context.setDefaultTimeout(10 * 1000) // 10s
+    const browser = proxy
+      ? await firefox.launch({
+          proxy: {
+            server: `${proxy.host}:${proxy.port}`,
+            username: proxy.username,
+            password: proxy.password,
+          },
+        })
+      : this.browser
+    const context = await browser.newContext({
+      extraHTTPHeaders: headers,
+    })
+    context.setDefaultTimeout(30000) // 30s
     if (blockAds) {
       context.route("**", (route) => {
         const url = new URL(route.request().url())
@@ -59,11 +57,37 @@ export class RendererService {
       }
       return `<!DOCTYPE HTML>${html}`
     } finally {
+      await context.close()
       if (proxy) {
-        proxyBrowser.close()
-      } else {
-        await context.close()
+        await browser.close()
       }
     }
+  }
+
+  public async renderSSR(
+    url: string,
+    headers?: Record<string, string>,
+    proxy?: HttpProxy
+  ): Promise<string> {
+    const res = await this.axios.get(url, {
+      headers,
+      proxy: proxy
+        ? {
+            host: proxy.host,
+            port: proxy.port,
+            auth:
+              proxy.username && proxy.password
+                ? {
+                    username: proxy.username,
+                    password: proxy.password,
+                  }
+                : undefined,
+          }
+        : undefined,
+    })
+    if (res.status < 200 || res.status >= 300 || !res.data) {
+      throw new Error(`Cannot extract HTML from ${url}`)
+    }
+    return res.data
   }
 }