+++ src/resources/resources.module.ts
@@ -1,11 +1,14 @@
 import { HttpModule, HttpService, Module } from "@nestjs/common"
+import { ConfigService } from "@nestjs/config"
 import { BlocklistProvider } from "./blocklist.provider"
 
 const providers = [
   {
     provide: BlocklistProvider.providerName,
-    inject: [HttpService],
-    useFactory: (httpService: HttpService) => BlocklistProvider.init(httpService),
+    inject: [ConfigService, HttpService],
+    useFactory: (config: ConfigService, http: HttpService) => {
+      return BlocklistProvider.init(config, http)
+    },
   },
 ]
 
@@ -14,4 +17,4 @@ const providers = [
   providers,
   exports: providers,
 })
-export class BlocklistModule {}
+export class ResourcesModule {}